vkplus/pyproject.toml
[tool.poetry]
name = "vkplus"
version = "2.1"
description = "VK selfbot written in Python using vkbottle"
authors = ["Timur"]
license = "GPL 3.0"

[tool.poetry.dependencies]
python = "^3.7.2"
vkbottle = "^4.0"
pillow = "^8.3.0"
aiohttp = "^3.6.2"
psutil = "^5.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
